move to shared util func
